Move fusion tx code into separate file
