+++ src/SimpleWallet/Fusion.h
@@ -1,0 +1,17 @@
+#include "CryptoNoteConfig.h"
+
+#include <SimpleWallet/Tools.h>
+
+#include <Wallet/WalletGreen.h>
+
+bool fusionTX(CryptoNote::WalletGreen &wallet, 
+              CryptoNote::TransactionParameters p);
+
+bool optimize(CryptoNote::WalletGreen &wallet, uint64_t threshold);
+
+void quickOptimize(CryptoNote::WalletGreen &wallet);
+
+void fullOptimize(CryptoNote::WalletGreen &wallet);
+
+size_t makeFusionTransaction(CryptoNote::WalletGreen &wallet, 
+                             uint64_t threshold);
